Adds endpoint and rejust the umbrellah project
