+++ apps/persistence/config/prod.exs
@@ -9,7 +9,8 @@ import Config
 # manifest is generated by the `mix phx.digest` task,
 # which you should run after static files are built and
 # before starting your production server.
-config :persistence, PersistenceWeb.Endpoint, cache_static_manifest: "priv/static/cache_manifest.json"
+config :persistence, PersistenceWeb.Endpoint,
+  cache_static_manifest: "priv/static/cache_manifest.json"
 
 # Do not print debug messages in production
 config :logger, level: :info
