+++ apps/producer/.formatter.exs
