+++ apps/producer/lib/producer/application.ex
@@ -1,0 +1,31 @@
+defmodule Producer.Application do
+  # See https://hexdocs.pm/elixir/Application.html
+  # for more information on OTP Applications
+  @moduledoc false
+
+  use Application
+
+  @impl true
+  def start(_type, _args) do
+    children = [
+      Producer.Telemetry,
+      # Start a worker by calling: Producer.Worker.start_link(arg)
+      # {Producer.Worker, arg},
+      # Start to serve requests, typically the last entry
+      Producer.Endpoint
+    ]
+
+    # See https://hexdocs.pm/elixir/Supervisor.html
+    # for other strategies and supported options
+    opts = [strategy: :one_for_one, name: Producer.Supervisor]
+    Supervisor.start_link(children, opts)
+  end
+
+  # Tell Phoenix to update the endpoint configuration
+  # whenever the application is updated.
+  @impl true
+  def config_change(changed, _new, removed) do
+    Producer.Endpoint.config_change(changed, removed)
+    :ok
+  end
+end
