+++ apps/producer/lib/producer/controllers/broadway_controller.ex
@@ -1,0 +1,35 @@
+defmodule Producer.BroadwayController do
+  use Producer, :controller
+  alias Broadway
+
+  plug :accepts, ~w(json ...)
+
+  def send_message(conn, params) do
+    message_build = build_params(params)
+
+    IO.inspect(params, label: :params)
+    IO.inspect(message_build, label: :message)
+
+    case Broadway.broadway_message(message_build) do
+      {:error, :unknow_reason} ->
+        render(conn, :index, %{erro: :unknow_reason})
+
+      {:error, body} ->
+        IO.inspect(body)
+        render(conn, :index, %{error: body})
+
+        {:ok, :message_send} ->
+        render(conn, :index, %{message: :sucess_send})
+    end
+  end
+
+  defp build_params(params) do
+    %{
+      endpoint: params["endpoint"],
+      client: params["client"],
+      currencie_from: params["currencie_from"],
+      currencie_to: params["currencie_to"],
+      value_to_convert: params["value_to_convert"]
+    }
+  end
+end
