+++ apps/producer/lib/producer/controllers/broadway_json.ex
@@ -1,0 +1,9 @@
+defmodule Producer.BroadwayJSON do
+  def index(%{message: :sucess_send}) do
+    "success send"
+  end
+
+  def index(%{error: reason}) do
+    reason
+  end
+end
