+++ apps/producer/mix.exs
@@ -1,13 +1,16 @@
-defmodule Persistence.MixProject do
+defmodule Producer.MixProject do
   use Mix.Project
 
   def project do
     [
-      app: :persistence,
+      app: :producer,
       version: "0.1.0",
-      elixir: "~> 1.12",
+      build_path: "../../_build",
+      config_path: "../../config/config.exs",
+      deps_path: "../../deps",
+      lockfile: "../../mix.lock",
+      elixir: "~> 1.14",
       elixirc_paths: elixirc_paths(Mix.env()),
-      compilers: Mix.compilers(),
       start_permanent: Mix.env() == :prod,
       aliases: aliases(),
       deps: deps()
@@ -19,7 +22,7 @@ defmodule Persistence.MixProject do
   # Type `mix help compile.app` for more information.
   def application do
     [
-      mod: {Persistence.Application, []},
+      mod: {Producer.Application, []},
       extra_applications: [:logger, :runtime_tools]
     ]
   end
@@ -35,26 +38,26 @@ defmodule Persistence.MixProject do
     [
       {:phoenix, "~> 1.7.12"},
       {:phoenix_ecto, "~> 4.4"},
-      {:ecto_sql, "~> 3.6"},
-      {:postgrex, ">= 0.0.0"},
+      {:phoenix_live_dashboard, "~> 0.8.3"},
       {:telemetry_metrics, "~> 1.0"},
       {:telemetry_poller, "~> 1.0"},
+      {:bandit, "~> 1.2"},
+      # HTTP client
+      {:tesla, "~> 1.4"},
+      {:hackney, "~> 1.17"},
       {:jason, "~> 1.2"},
-      {:plug_cowboy, "~> 2.5"}
+      # Kafka connection
+      {:kaffe, "~> 1.0"},
+      {:persistence, in_umbrella: true}
     ]
   end
 
   # Aliases are shortcuts or tasks specific to the current project.
-  # For example, to install project dependencies and perform other setup tasks, run:
-  #
-  #     $ mix setup
   #
   # See the documentation for `Mix` for more info on aliases.
   defp aliases do
     [
-      setup: ["deps.get", "ecto.setup"],
-      "ecto.setup": ["ecto.create", "ecto.migrate", "run priv/repo/seeds.exs"],
-      "ecto.reset": ["ecto.drop", "ecto.setup"],
+      setup: ["deps.get"],
       test: ["ecto.create --quiet", "ecto.migrate --quiet", "test"]
     ]
   end
