+++ apps/producer/test/producer/controllers/error_json_test.exs
@@ -1,0 +1,12 @@
+defmodule Producer.ErrorJSONTest do
+  use Producer.ConnCase, async: true
+
+  test "renders 404" do
+    assert Producer.ErrorJSON.render("404.json", %{}) == %{errors: %{detail: "Not Found"}}
+  end
+
+  test "renders 500" do
+    assert Producer.ErrorJSON.render("500.json", %{}) ==
+             %{errors: %{detail: "Internal Server Error"}}
+  end
+end
