+++ apps/producer/test/test_helper.exs
@@ -1,0 +1,2 @@
+ExUnit.start()
+Ecto.Adapters.SQL.Sandbox.mode(Producer.Repo, :manual)
