--- apps/producer_client/lib/producer_client/application.ex
@@ -1,34 +1,0 @@
-defmodule ProducerClient.Application do
-  # See https://hexdocs.pm/elixir/Application.html
-  # for more information on OTP Applications
-  @moduledoc false
-
-  use Application
-
-  @impl true
-  def start(_type, _args) do
-    children = [
-      # Start the Telemetry supervisor
-      ProducerClientWeb.Telemetry,
-      # Start the PubSub system
-      {Phoenix.PubSub, name: ProducerClient.PubSub},
-      # Start the Endpoint (http/https)
-      ProducerClientWeb.Endpoint
-      # Start a worker by calling: ProducerClient.Worker.start_link(arg)
-      # {ProducerClient.Worker, arg}
-    ]
-
-    # See https://hexdocs.pm/elixir/Supervisor.html
-    # for other strategies and supported options
-    opts = [strategy: :one_for_one, name: ProducerClient.Supervisor]
-    Supervisor.start_link(children, opts)
-  end
-
-  # Tell Phoenix to update the endpoint configuration
-  # whenever the application is updated.
-  @impl true
-  def config_change(changed, _new, removed) do
-    ProducerClientWeb.Endpoint.config_change(changed, removed)
-    :ok
-  end
-end
