--- apps/producer_client/lib/producer_client_web/views/broadway_view.ex
@@ -1,7 +1,0 @@
-defmodule ProducerClientWeb.BroadwayView do
-  use ProducerClientWeb, :view
-
-  def show(_conn, "show.json", params) do
-    params
-  end
-end
