--- apps/producer_client/test/producer_client_web/views/error_view_test.exs
@@ -1,15 +1,0 @@
-defmodule ProducerClientWeb.ErrorViewTest do
-  use ProducerClientWeb.ConnCase, async: true
-
-  # Bring render/3 and render_to_string/3 for testing custom views
-  import Phoenix.View
-
-  test "renders 404.json" do
-    assert render(ProducerClientWeb.ErrorView, "404.json", []) == %{errors: %{detail: "Not Found"}}
-  end
-
-  test "renders 500.json" do
-    assert render(ProducerClientWeb.ErrorView, "500.json", []) ==
-             %{errors: %{detail: "Internal Server Error"}}
-  end
-end
