+++ config/config.exs
@@ -9,6 +9,21 @@
 # move said applications out of the umbrella.
 import Config
 
+config :producer,
+  ecto_repos: [Producer.Repo],
+  generators: [context_app: false]
+
+# Configures the endpoint
+config :producer, Producer.Endpoint,
+  url: [host: "localhost"],
+  adapter: Bandit.PhoenixAdapter,
+  render_errors: [
+    formats: [json: Producer.ErrorJSON],
+    layout: false
+  ],
+  pubsub_server: Producer.PubSub,
+  live_view: [signing_salt: "VCY3c1oQ"]
+
 # Sample configuration:
 #
 #     config :logger, :console,
@@ -21,4 +36,19 @@ config :phoenix, :json_library, Jason
 config :kaffe,
   producer: [
     endpoints: [localhost: 9092],
-    topics: ["currencie_converter"]]
+    topics: ["currencie_converter"]
+  ]
+
+import Config
+
+# Configures Elixir's Logger
+config :logger, :console,
+  format: "$time $metadata[$level] $message\n",
+  metadata: [:request_id]
+
+# Use Jason for JSON parsing in Phoenix
+config :phoenix, :json_library, Jason
+
+# Import environment specific config. This must remain at the bottom
+# of this file so it overrides the configuration defined above.
+import_config "#{config_env()}.exs"
